pyproject.toml
[project]
name = "agentich-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "langchain>=0.2",
    "langchain-community>=0.2",
    "langchain-core>=0.2",
    "langchain-groq>=0.1.5",
    "langgraph>=0.2",
    "python-dotenv>=1.0",
    "tavily-python>=0.3",
    "typing-extensions>=4.9",
    "uvicorn>=0.27",
]
